systemsim_idle: update to be usable as a ppc_md.power_save() function
In kernel 3.8 the idle_loop() function pointer in struct machdep_calls was
removed (commit 16b86bf2520ed29712ca7462dbfe76c856b445e9) as the only user
of it was the legacy iSeries support which was removed earlier.

So update the systemsim_idle() function so that it can be used as a
power_save() function. See commit 302eca184fb844670fb128c69e22a8a28bbce48a
([POWERPC] cell: use ppc_md->power_save instead of cbe_idle_loop) for an
example of this same transformation but in this case for the Cell idle code.

Signed-off-by: Mark Nelson <mdnelson8@gmail.com>